Improved overall design

+++ base_app/views.py
@@ -59,8 +59,9 @@ def logoutUser(request):
 
 
 def home(request):
-    categories = models.Category.objects.all()
-    context = {"categories": categories}
+    categories = [category for category in models.Category.objects.all()]
+    last_category = categories.pop()
+    context = {"categories": categories, "last_category": last_category}
     return render(request, "home.html", context)
 
 
@@ -73,14 +74,18 @@ def category(request, pk):
 
 def results(request):
     questions = []
-    category = request.POST.get("category")
-    category = models.Category.objects.get(name=category)
+    try:
+        category = request.POST.get("category")
+        category = models.Category.objects.get(name=category)
+    except:
+        return redirect('home')
     questions = category.question_set.all()
 
     class ResponseSet:
-        def __init__(self, question, answer):
-            self.a = answer
-            if self.a == question.answer:
+        def __init__(self, question, chosen):
+            self.chosen = chosen
+            self.question = question
+            if self.chosen == question.answer:
                 self.is_correct = True
             else:
                 self.is_correct = False
@@ -90,6 +95,28 @@ def results(request):
                 self.points = question.points
             else:
                 self.points = 0
+
+            if chosen == "a":
+                self.supp = question.a
+            elif chosen == "b":
+                self.supp = question.b
+            elif chosen == "c":
+                self.supp = question.c
+            elif chosen == "d":
+                self.supp = question.d
+            elif chosen == "e":
+                self.supp = question.e
+
+            if question.answer == "a":
+                self.correct_choice = question.a
+            elif question.answer == "b":
+                self.correct_choice = question.b
+            elif question.answer == "c":
+                self.correct_choice = question.c
+            elif question.answer == "d":
+                self.correct_choice = question.d
+            elif question.answer == "d":
+                self.correct_choice = question.e
 
     response_sets = []
     for question in questions:
@@ -168,6 +195,7 @@ def view_high_scores(request, pk):
             fin.append((final_sorted[i], tup[i]))
         return fin
 
+    high_scores = list(set(high_scores))
     high_scores = sort_tuples(high_scores)
     pseudo_scores = []
     for score_set in high_scores:
@@ -176,13 +204,32 @@ def view_high_scores(request, pk):
     context = {
         "pseudo_scores": pseudo_scores,
         "category": category,
-        "high_scores": high_scores,
+        "count": len(pseudo_scores),
+        "name": category.name.upper(),
     }
     return render(request, "high_scores.html", context)
 
 
 def inter(request):
-    categories = models.Category.objects.all()
+    if not request.user.is_authenticated:
+        return redirect("home")
+    categories1 = models.Category.objects.all()
+    categories = []
+
+    class High:
+        def __init__(self, category):
+            self.name = category.name
+            if len([_ for _ in category.userresults_set.all()]) > 0:
+                self.highest_score = max(
+                    [result.score for result in category.userresults_set.all()]
+                )
+            else:
+                self.highest_score = 0
+            self.id = category.id
+
+    for category in categories1:
+        # if len([ _ for _ in category.userresults_set.all()]) > 0:
+        categories.append(High(category))
     return render(request, "inter_high.html", {"categories": categories})
 
 
@@ -190,6 +237,9 @@ def admin_panel(request):
     if not request.user.is_superuser:
         return redirect("home")
     categories = models.Category.objects.all()
+
+    if request.method == "POST" and request.POST.get("name") != "":
+        models.Category.objects.create(name=request.POST.get("name"))
 
     class PseudoCategory:
         def __init__(self, category):
@@ -205,6 +255,8 @@ def admin_panel(request):
 
 
 def delete_question(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
     question = models.Question.objects.get(id=pk)
     if request.method == "POST":
         question.delete()
@@ -213,6 +265,8 @@ def delete_question(request, pk):
 
 
 def add_to_category(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
     category = models.Category.objects.get(id=pk)
     form = forms.QuestionForm()
     if request.method == "POST":
@@ -233,6 +287,8 @@ def add_to_category(request, pk):
 
 
 def edit_question(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
     question = models.Question.objects.get(id=pk)
     form = forms.QuestionForm(instance=question)
     if request.method == "POST":
@@ -247,3 +303,35 @@ def edit_question(request, pk):
         question.save()
         return redirect("admin_panel")
     return render(request, "edit_question.html", {"form": form})
+
+
+def delete_category(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
+    category = models.Category.objects.get(id=pk)
+    if request.method == "POST":
+        category.delete()
+        return redirect("admin_panel")
+    return render(request, "delete.html", {"obj": "category"})
+
+
+def edit_category(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
+    category = models.Category.objects.get(id=pk)
+    if request.method == "POST":
+        category.name = request.POST.get("name")
+        category.save()
+        return redirect("admin_panel")
+    return render(request, "edit_category.html", {"category": category})
+
+
+def reset_highscores(request, pk):
+    if not request.user.is_superuser:
+        return redirect("home")
+    category = models.Category.objects.get(id=pk)
+    if request.method == "POST":
+        for result in category.userresults_set.all():
+            result.delete()
+        return redirect("view_high_scores", pk)
+    return render(request, "reset.html", {"name": category.name})
